test(control-flow): isolate AERROR SQL/OLE metadata fixture
