+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 - 2026-08-21: Continued #2564 test-module refactoring by moving the focused
+  SQL/OLE `AERROR()` metadata fixture into
+  `test_prg_engine_control_flow_aerror_sql_ole_metadata.cpp`. The existing
+  executable, invocation order, assertions, and runtime behavior remain
+  unchanged.
+
+- 2026-08-21: Continued #2564 test-module refactoring by moving the focused
   `AERROR()` runtime-metadata fixture into
   `test_prg_engine_control_flow_aerror_runtime_metadata.cpp`. The existing
   executable, invocation order, assertions, and runtime behavior remain
